Create the_cake_is_not_a_lie
One of Google's Foo.Bar challenges, the solution here was one of several attempts against the test cases provided but due to an unknown issue these continued to fail in browser. This is not the most elegant nor the final solution i wanted to submit but it is where i stopped work within the time limit so i will record it as is. Test cases did fortunately work on the desktop so there was progress.